Move the default path implementation into a module named default.
