+++ examples/gfx_advanced/src/main.rs
@@ -13,7 +13,7 @@ use lyon::tessellation::basic_shapes::*;
 use lyon::tessellation::{FillTessellator, FillOptions};
 use lyon::tessellation::{StrokeTessellator, StrokeOptions};
 use lyon::tessellation;
-use lyon::path::Path;
+use lyon::path::default::Path;
 
 use gfx::traits::{Device, FactoryExt};
