+++ path/src/lib.rs
@@ -11,7 +11,7 @@
 //! ```
 //! # extern crate lyon_path;
 //! # fn main() {
-//! use lyon_path::Path;
+//! use lyon_path::default::Path;
 //! use lyon_path::math::{point};
 //! use lyon_path::builder::*;
 //!
@@ -38,13 +38,12 @@
 
 pub extern crate lyon_geom as geom;
 
-mod path;
 mod events;
 mod path_state;
+pub mod default;
 pub mod iterator;
 pub mod builder;
 
-pub use path::*;
 pub use events::*;
 pub use path_state::*;
 pub use geom::ArcFlags;
