+++ tessellation/src/fuzz_tests.rs
@@ -1,5 +1,5 @@
 use geometry_builder::{VertexBuffers, simple_builder};
-use path::{Path, PathSlice};
+use path::default::{Path, PathSlice};
 use path::builder::FlatPathBuilder;
 use path_fill::*;
 use geom::math::*;
