+++ tessellation/src/path_stroke.rs
@@ -42,7 +42,7 @@ use std::f32::consts::PI;
 ///
 /// ```
 /// # extern crate lyon_tessellation as tess;
-/// # use tess::path::Path;
+/// # use tess::path::default::Path;
 /// # use tess::path::builder::*;
 /// # use tess::path::iterator::*;
 /// # use tess::geom::math::*;
@@ -936,7 +936,7 @@ fn tess_round_cap(
 }
 
 #[cfg(test)]
-use path::{Path, PathSlice};
+use path::default::{Path, PathSlice};
 #[cfg(test)]
 use geometry_builder::{
     SimpleBuffersBuilder, simple_builder, VertexBuffers,
